MDL-70658 phpunit: never apply for time() in data providers
Data providers are executed at the beginning of the execution
of the phpunit run, not when the test is run.

So, setting anything on them being dependent of current time
has high chances or fail when the test is executed (maybe some
hours later in slow databases / systems).

Data provider only can contain the relative values and then,
when the test is effectively run, they are applied to *current* time.