MDL-73357 mod_data: Update behat tests
After replacing the tabs navigation with the secondary and tertiary
navigation, the completion data will be no longer displayed on all
pages in the database activity. Therefore, we need to update the
behat tests that still verify the existance of completion data in
pages other than the database view page.