MDL-75751 upgradelib: Refactor out duplicated component update code
There are many places (install/upgrade for core, modules, blocks,
and plugins) where we need to call the same, or very similar,
sequence of Moodle functions to ensure the item is updated properly.

This change moves that sequence into a function so that it can be
held consistently in one place.