MDL-78934 behat: Move from mink goutte to browserkit driver
In order to get compatibility with PHP 8.1 there are 2 options:

1) Move to the Goutte 2 mink driver, that uses Goutte 4.
2) Move to the BrowserKit mink driver.

While 1) could sound like the easier alternative, and that is
what we are doing right now, there are a few points to take into consideration.

a) Goutte 4 is a complete wrapper over BrowserKit. 100%. See its
   source code: https://github.com/FriendsOfPHP/Goutte/blob/master/Goutte/Client.php
b) The mink extension still doesn't support Goutte 4, but supports
   BrowserKit. See https://github.com/FriendsOfBehat/MinkExtension (PR#16)
   (we could workaround this by forking the package again)
c) Last, but not least, recently (June 2023) both the MinkGoutteDriver
   and Goutte itself have been archived/abandoned:
   - https://packagist.org/packages/behat/mink-goutte-driver
   - https://packagist.org/packages/fabpot/goutte

So, given the points above, this is an attempt to move away
completely from Goutte and straight use the BrowserKit alternative
that, in practice, is the very same code that we are already using
(you can see that, in the .lock changes there isn't any new thing
being installed, because, as said, we are using it already).

Let's see how complex it is, I imagine it won't be hard. Surely killing
all references everywhere about Goutte will be the most complex task.